api/requirements.txt
fastapi==0.109.2
sqlalchemy==2.0.27
python-dotenv==1.0.0
asyncpg==0.29.0
uvicorn[standard]==0.27.1